+++ src/classes/index.ts
@@ -1,2 +1,3 @@
+export * from "./DefaultPermissionResolver";
 export * from "./SimpleCommandMessage";
 export * from "./errors";