+++ src/decorators/classes/DApplicationCommand.ts
@@ -8,6 +8,7 @@ import {
 
 import {
   DApplicationCommandOption,
+  IDefaultPermission,
   IGuild,
   IPermissions,
   resolveIPermission,
@@ -21,7 +22,7 @@ export class DApplicationCommand extends Method {
   private _name: string;
   private _description: string;
   private _type: ApplicationCommandType;
-  private _defaultPermission: boolean;
+  private _defaultPermission: IDefaultPermission;
   private _options: DApplicationCommandOption[] = [];
   private _permissions: IPermissions[] = [];
   private _guilds: IGuild[];
@@ -71,10 +72,10 @@ export class DApplicationCommand extends Method {
     this._guilds = value;
   }
 
-  get defaultPermission(): boolean {
+  get defaultPermission(): IDefaultPermission {
     return this._defaultPermission;
   }
-  set defaultPermission(value: boolean) {
+  set defaultPermission(value: IDefaultPermission) {
     this._defaultPermission = value;
   }
 
@@ -151,14 +152,20 @@ export class DApplicationCommand extends Method {
     return option;
   }
 
-  toJSON(config?: { channelString: boolean }): ApplicationCommandData {
+  async toJSON(config?: {
+    channelString?: boolean;
+    guild?: Guild;
+  }): Promise<ApplicationCommandData> {
     const options = [...this.options]
       .reverse()
       .map((option) => option.toJSON(config));
 
     if (this.type === "CHAT_INPUT") {
       return {
-        defaultPermission: this.defaultPermission,
+        defaultPermission:
+          typeof this.defaultPermission === "boolean"
+            ? this.defaultPermission
+            : await this.defaultPermission.resolver(config?.guild),
         description: this.description,
         name: this.name,
         options: options,
@@ -167,7 +174,10 @@ export class DApplicationCommand extends Method {
     }
 
     return {
-      defaultPermission: this.defaultPermission,
+      defaultPermission:
+        typeof this.defaultPermission === "boolean"
+          ? this.defaultPermission
+          : await this.defaultPermission.resolver(config?.guild),
       description: "",
       name: this.name,
       options: [],