+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - throw error for duplicate simple command
+- resolver for default permission
 
 ## [7.0.0] - 2021-10-20